refactor email_functions
